requirements.txt
colorlog==4.0.2
cycler==0.10.0
decorator==4.4.0
ipython==6.5.0
ipython-genutils==0.2.0
matplotlib==3.0.0
numpy==1.15.2
olefile==0.46
packaging==17.2
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
protobuf==3.6.0
Pygments==2.2.0
pyparsing==2.4.0
python-dateutil==2.7.3
pytz==2019.1
scikit-learn==0.20.0
scipy==1.1.0
simplegeneric==0.8.1
six==1.10.0
tensorflow-gpu==1.10.0
termcolor==1.1.0
tqdm==4.31.1
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
